fix: address responses review feedback
