+++ lmdeploy/serve/openai/responses/__init__.py
@@ -13,17 +13,19 @@ from .protocol import (
     ResponsesResponse,
     ResponseUsage,
 )
-from .serving import (
-    _generation_messages_from_parser,
-    _make_response,
+from .request import (
     _messages_from_input,
     _openai_tools_from_responses,
-    _stream_response,
     _to_generation_config,
     _tool_choice_from_responses,
     _validate_text_v1_request,
+)
+from .response import _make_response
+from .serving import (
+    OpenAIServingResponses,
     create_responses_router,
 )
+from .streaming import _stream_response
 
 __all__ = [
     'ResponseIncompleteDetails',
@@ -36,8 +38,8 @@ __all__ = [
     'ResponsesRequest',
     'ResponsesResponse',
     'ResponseUsage',
+    'OpenAIServingResponses',
     'create_responses_router',
-    '_generation_messages_from_parser',
     '_make_response',
     '_messages_from_input',
     '_openai_tools_from_responses',
